Goodread_5_api_2/djangoProject/requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.2
django-crispy-forms==2.0
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
